+++ src/main/kotlin/online/bingzi/luck/perms/bridge/spring/boot/starter/entity/request/QueryMode.kt
@@ -1,0 +1,11 @@
+package online.bingzi.luck.perms.bridge.spring.boot.starter.entity.request
+
+/**
+ * 查询模式
+ */
+enum class QueryMode {
+    /** 上下文相关 */
+    CONTEXTUAL,
+    /** 上下文无关 */
+    NON_CONTEXTUAL
+}